[CodingStandard] Add AbstractManyNodeTypeRule
